# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Courier & Logistics API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface TrackingEvent {
  timestamp: string;
  location: string;
  status: string;
  description: string;
}

export interface TrackingResult {
  trackingNumber: string;
  status: string;
  /** @nullable */
  estimatedDelivery?: string | null;
  origin: string;
  destination: string;
  service: string;
  /** @nullable */
  weight?: string | null;
  events: TrackingEvent[];
}

export interface Shipment {
  id: number;
  trackingNumber: string;
  status: string;
  service: string;
  origin: string;
  destination: string;
  /** @nullable */
  estimatedDelivery?: string | null;
  /** @nullable */
  weight?: string | null;
  /** @nullable */
  cost?: number | null;
  createdAt: string;
  insured: boolean;
}

export interface ShipmentInput {
  service: string;
  origin: string;
  destination: string;
  weight: string;
  insured?: boolean;
}

export interface DeliveryService {
  id: string;
  name: string;
  description: string;
  deliveryTime: string;
  priceFrom: number;
  category: string;
  features: string[];
}

export interface ContactInput {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
}

export interface ContactResponse {
  success: boolean;
  message: string;
}

export interface PickupInput {
  name: string;
  address: string;
  city: string;
  date: string;
  timeSlot: string;
  phone: string;
  /** @nullable */
  notes?: string | null;
}

export interface PickupResponse {
  success: boolean;
  confirmationCode: string;
  message: string;
}

export interface DashboardStats {
  totalShipments: number;
  activeShipments: number;
  deliveredThisMonth: number;
  pendingPickups: number;
  totalSpent: number;
}