# lib/api-zod/src/generated/types/deliveryService.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Courier & Logistics API
 * OpenAPI spec version: 0.1.0
 */

export interface DeliveryService {
  id: string;
  name: string;
  description: string;
  deliveryTime: string;
  priceFrom: number;
  category: string;
  features: string[];
}