# lib/api-zod/src/generated/types/pickupInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Courier & Logistics API
 * OpenAPI spec version: 0.1.0
 */

export interface PickupInput {
  name: string;
  address: string;
  city: string;
  date: string;
  timeSlot: string;
  phone: string;
  /** @nullable */
  notes?: string | null;
}