# lib/api-zod/src/generated/types/pickupResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Courier & Logistics API
 * OpenAPI spec version: 0.1.0
 */

export interface PickupResponse {
  success: boolean;
  confirmationCode: string;
  message: string;
}