# lib/api-zod/src/generated/types/shipment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Courier & Logistics API
 * OpenAPI spec version: 0.1.0
 */

export interface Shipment {
  id: number;
  trackingNumber: string;
  status: string;
  service: string;
  origin: string;
  destination: string;
  /** @nullable */
  estimatedDelivery?: string | null;
  /** @nullable */
  weight?: string | null;
  /** @nullable */
  cost?: number | null;
  createdAt: string;
  insured: boolean;
}