# lib/api-zod/src/generated/types/shipmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Courier & Logistics API
 * OpenAPI spec version: 0.1.0
 */

export interface ShipmentInput {
  service: string;
  origin: string;
  destination: string;
  weight: string;
  insured?: boolean;
}