# lib/api-zod/src/generated/types/trackingEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Courier & Logistics API
 * OpenAPI spec version: 0.1.0
 */

export interface TrackingEvent {
  timestamp: string;
  location: string;
  status: string;
  description: string;
}