# lib/api-zod/src/generated/types/trackingResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Courier & Logistics API
 * OpenAPI spec version: 0.1.0
 */
import type { TrackingEvent } from './trackingEvent';

export interface TrackingResult {
  trackingNumber: string;
  status: string;
  /** @nullable */
  estimatedDelivery?: string | null;
  origin: string;
  destination: string;
  service: string;
  /** @nullable */
  weight?: string | null;
  events: TrackingEvent[];
}